App/requirements.txt
alembic==1.10.4
blinker==1.6.2
click==8.1.3
colorama==0.4.6
Flask==2.3.2
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
importlib-metadata==6.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
PyMySQL==1.0.3
python-dotenv==1.0.0
pytz==2023.3
SQLAlchemy==2.0.12
typing_extensions==4.5.0
Werkzeug==2.3.3
zipp==3.15.0
